LayoutTests/crypto/workers/subtle/hkdf-import-key-derive-hmac-key-expected.txt

[Worker] Test deriving HMAC Keys with imported HKDF base key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/hkdf-import-key-derive-hmac-key.js
PASS [Worker] derivedKey.type is 'secret'
PASS [Worker] derivedKey.extractable is false
PASS [Worker] derivedKey.algorithm.name is 'HMAC'
PASS [Worker] derivedKey.algorithm.length is 512
PASS [Worker] derivedKey.algorithm.hash.name is 'SHA-1'
PASS [Worker] derivedKey.usages is ['sign', 'verify']
PASS successfullyParsed is true

TEST COMPLETE
